Move session setup to extension method
