+++ CopticChanter.WebApi/Program.cs
@@ -1,7 +1,5 @@
-using CopticChanter.WebApi;
-using CoptLib.IO;
+using CopticChanter.WebApi.DependencyInjection;
 using CoptLib.Scripting;
-using System.Collections.Concurrent;
 
 const string allowedOrigins = nameof(allowedOrigins);
 
@@ -28,24 +26,7 @@ builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 builder.Services.AddControllers();
 
-builder.Services.AddKeyedSingleton<ConcurrentDictionary<string, Session>>("sessions");
-builder.Services.AddScoped(ctx =>
-{
-    var contextAccessor = ctx.GetRequiredService<IHttpContextAccessor>();
-    var httpContext = contextAccessor.HttpContext!;
-
-    var sessionKey = httpContext.Request.Query["sessionKey"].FirstOrDefault()
-        ?? Guid.NewGuid().ToString("N").ToUpperInvariant();
-
-    var sessions = ctx.GetRequiredKeyedService<ConcurrentDictionary<string, Session>>("sessions");
-    var context = sessions.GetOrAdd(sessionKey, k =>
-    {
-        ILoadContext context = new LoadContext();
-        var env = ctx.GetRequiredService<IWebHostEnvironment>();
-        return new(k, context, env);
-    });
-    return context;
-});
+builder.Services.AddCoptSessions();
 
 DotNetScript.Register();
 LuaScript.Register();
